+++ CHANGELOG.md
@@ -14,9 +14,10 @@
 <hr/>
 
 ## 5.1.1 (Sep 22, 2019)
-### :bug: Bug Fix
+### :rocket: New Feature
 - `image`
   - [#388](https://github.com/wix-incubator/rich-content/pull/388) changed default image alignment
+### :bug: Bug Fix
 - `viewer`
   - [#390](https://github.com/wix-incubator/rich-content/pull/390) justified text new lines disappear
 ### :house: Internal